+++ src/main/java/com/example/flab/soft/shoppingmallfashion/sms/CoolSmsService.java
@@ -1,0 +1,34 @@
+package com.example.flab.soft.shoppingmallfashion.sms;
+
+import jakarta.annotation.PostConstruct;
+import net.nurigo.sdk.NurigoApp;
+import net.nurigo.sdk.message.model.Message;
+import net.nurigo.sdk.message.request.SingleMessageSendingRequest;
+import net.nurigo.sdk.message.service.DefaultMessageService;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+
+@Service
+public class CoolSmsService implements SmsService{
+    @Value("${coolsms.api.key}")
+    private String apiKey;
+    @Value("${coolsms.api.secret}")
+    private String apiSecretKey;
+    @Value("${coolsms.api.number}")
+    private String from;
+    private DefaultMessageService messageService;
+
+    @PostConstruct
+    private void init(){
+        this.messageService = NurigoApp.INSTANCE.initialize(apiKey, apiSecretKey, "https://api.coolsms.co.kr");
+    }
+
+    @Override
+    public void sendOne(String to, String code) {
+        Message message = new Message();
+        message.setFrom(from);
+        message.setTo(to);
+        message.setText("[MallFashion] 아래의 인증번호를 입력해주세요\n" + code);
+        messageService.sendOne(new SingleMessageSendingRequest(message));
+    }
+}
