+++ src/main/java/com/example/flab/soft/shoppingmallfashion/sms/SendCodeRequest.java
@@ -1,8 +1,16 @@
 package com.example.flab.soft.shoppingmallfashion.sms;
 
+import jakarta.validation.constraints.Pattern;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
 public class SendCodeRequest {
-    private String number;
+    @Pattern(regexp = "^010[0-9]{8}$")
+    private String phoneNumber;
 }
