+++ src/main/java/com/example/flab/soft/shoppingmallfashion/sms/SmsService.java
@@ -1,0 +1,5 @@
+package com.example.flab.soft.shoppingmallfashion.sms;
+
+public interface SmsService {
+    void sendOne(String to, String code);
+}
