+++ src/main/java/com/example/flab/soft/shoppingmallfashion/sms/VerifyCodeRequest.java
@@ -1,9 +1,18 @@
 package com.example.flab.soft.shoppingmallfashion.sms;
 
+import jakarta.validation.constraints.Pattern;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
 public class VerifyCodeRequest {
-    private String number;
+    @Pattern(regexp = "^010[0-9]{8}$")
+    private String phoneNumber;
+    @Pattern(regexp = "^[1-9]\\d{3}$")
     private String code;
 }
